custom hook for form instantiation
